skills/codon-optimization/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codon-optimization"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "python-codon-tables>=0.1.12",
    "dnachisel>=3.2.11",
    "biopython>=1.83",
    "click>=8.0",
    "pytest>=7.0",
    "pyrodigal>=3.0",
]

[project.scripts]
codonopt = "codonopt.cli:cli"
